Add an optional IP address whitelist for tutors
The whitelist is disabled by default and does not affect site admins. It
can be configured through the admin interface.

It uses the postgresql inet support for storing hosts and nets and for
figuring out, if a host is part of a net. This works for both IPv4 and
IPv6.

+++ templates/static/admin-tutors.js
@@ -1,9 +1,14 @@
 document.addEventListener("DOMContentLoaded", () => {
     document.querySelector("#add-tutor")
         .addEventListener("submit", onNewTutor);
+    document.querySelector("#add-ip-whitelist")
+        .addEventListener("submit", onNewIpWhitelistEntry);
 
     for(let removeButton of document.querySelectorAll("#add-tutor .button.remove")) {
         removeButton.addEventListener("click", onDeleteTutor);
+    }
+    for(let removeButton of document.querySelectorAll("#add-ip-whitelist .button.remove")) {
+        removeButton.addEventListener("click", onDeleteIpWhitelistEntry);
     }
 
     for(let adminCheckbox of document.querySelectorAll("#add-tutor input.admin")) {
@@ -44,6 +49,38 @@ async function onNewTutor(event) {
     }
 }
 
+async function onNewIpWhitelistEntry(event) {
+    event.preventDefault();
+
+    let ipnet = document.querySelector("#add-ip-whitelist input[name='ipnet']").value.trim();
+    if(ipnet.length == 0) {
+        toast("error", "Ungültige Eingabe");
+        return;
+    }
+    let year = parseInt(document.body.dataset.year);
+
+    try {
+        let url = "/api/ip-whitelist";
+
+        let response = await myfetch(url, {
+            method: "POST",
+            headers: new Headers({"Content-Type": "application/json"}),
+            body: JSON.stringify({
+                ipnet: ipnet,
+                year: year,
+            })
+        });
+        handleResponse(response, {
+            500: "Serverfehler (eventuell falsche IP-Syntax?)"
+        });
+
+        // reload to avoid rendering on the client
+        location.reload(true);
+    } catch(e) {
+        toast("error", e);
+    }
+}
+
 async function onDeleteTutor(event) {
     let targetRow = event.target.closest("tr");
 
@@ -56,6 +93,30 @@ async function onDeleteTutor(event) {
 
     try {
         let url = "/api/tutor/" + id;
+
+        let response = await myfetch(url, {
+            method: "DELETE"
+        });
+        handleResponse(response);
+
+        targetRow.parentNode.removeChild(targetRow);
+    } catch(e) {
+        toast("error", e);
+    }
+}
+
+async function onDeleteIpWhitelistEntry(event) {
+    let targetRow = event.target.closest("tr");
+
+    let id = targetRow.dataset.id;
+    let name = targetRow.querySelector("td:nth-of-type(1)").textContent;
+
+    if(!confirm(name + " wirklich löschen?")) {
+        return;
+    }
+
+    try {
+        let url = "/api/ip-whitelist/" + id;
 
         let response = await myfetch(url, {
             method: "DELETE"
